add mail send aop sample
